remember to change the font.family and keep with it
It also knows whether your font is serif or sans-serif. If you try it with serif and it doesn’t work, change both of them to sans-serif instead. For example, if I changed my font to Georgia down below I’d need it to be serif instead, since it’s a font with serifs.
http://jonathansoma.com/lede/data-studio/matplotlib/changing-fonts-in-matplotlib/